requirements.txt
flask 
google-cloud-storage
google-generativeai
google-cloud-aiplatform
google-auth
google-auth-oauthlib
google-auth-httplib2
gunicorn